Replace shapeless parent remap with proxy mesh and unify proxy API
Instead of remapping shapeless parent bodies to their nearest
shape-bearing ancestor (which broke joint pivot coordinates by
interpreting local-frame pivots in the wrong reference frame), create
lightweight 1-vertex ABD proxies at each shapeless body's world pose.
Newton FIXED joints already in the model naturally constrain the proxy
to its neighbour, eliminating the need for explicit ancestor lookup.

Key changes:
- Remove _compute_shape_body_anchors and the remap/effective_parent_body
  logic in favour of a pre-pass that creates proxies for all shapeless
  bodies referenced by joints
- Unify _get_or_create_anchor and _get_or_create_shapeless_proxy into a
  single _create_proxy(name, transform, *, is_fixed) API with a shared
  _proxy_slots cache; world-anchored joints share one "world_anchor"
  proxy
- Add _validate_ball_anchors for world-space anchor coincidence checks
  on spherical joints, matching the existing revolute/prismatic
  validators
- Refactor build_body_mesh into a standalone function, replace
  np.float64 casts with float(), and clean up type annotations

+++ newton/_src/solvers/uipc/cloth.py
@@ -162,7 +162,7 @@ class ClothBuilder:
         if model.tri_count > 0 and model.tri_materials is not None:
             tri_materials_np = model.tri_materials.numpy()
             # tri_materials layout: (tri_count, 5) -> [ke, ka, kd, drag, lift]
-            avg_ke = np.float64(np.mean(tri_materials_np[:, 0]))
+            avg_ke = float(np.mean(tri_materials_np[:, 0]))
             if avg_ke > 0:
                 return avg_ke
         return 1000.0  # Default: 1 kPa
@@ -172,7 +172,7 @@ class ClothBuilder:
         if model.edge_count > 0 and model.edge_bending_properties is not None:
             edge_props_np = model.edge_bending_properties.numpy()
             # edge_bending_properties layout: (edge_count, 2) -> [stiffness, damping]
-            avg_ke = np.float64(np.mean(edge_props_np[:, 0]))
+            avg_ke = float(np.mean(edge_props_np[:, 0]))
             if avg_ke > 0:
                 return avg_ke
         return self._default_bending_stiffness
@@ -184,9 +184,9 @@ class ClothBuilder:
         """
         if model.particle_mass is not None:
             particle_mass_np = model.particle_mass.numpy()
-            total_mass = np.float64(np.sum(particle_mass_np[particle_indices]))
+            total_mass = float(np.sum(particle_mass_np[particle_indices]))
             if total_mass > 0 and model.tri_count > 0 and model.tri_areas is not None:
-                total_area = np.float64(np.sum(model.tri_areas.numpy()))
+                total_area = float(np.sum(model.tri_areas.numpy()))
                 if total_area > 0:
                     # Surface density = total_mass / total_area
                     # Volume density = surface_density / thickness
